Reorganize/rename autoqa scripts
