+++ tool/autogen-annotations.js
@@ -15,7 +15,7 @@ const util = require('util');
 const ThingTalk = require('thingtalk');
 
 const { parseConstantFile } = require('./lib/constant-file');
-const Annotator = require('./lib/webqa-canonical-annotator');
+const CanonicalGenerator = require('./lib/canonical-generator');
 const StreamUtils = require('../lib/stream-utils');
 
 async function loadClassDef(thingpedia) {
@@ -26,7 +26,7 @@ async function loadClassDef(thingpedia) {
 
 module.exports = {
     initArgparse(subparsers) {
-        const parser = subparsers.addParser('webqa-update-canonicals', {
+        const parser = subparsers.addParser('autogen-annotations', {
             addHelp: true,
             description: "Use BERT to expand canonicals"
         });
@@ -115,8 +115,8 @@ module.exports = {
         } else {
             const options = args;
             const constants = await parseConstantFile(args.locale, args.constants);
-            const generator = new Annotator(classDef, constants, args.queries.split(','), args.parameter_datasets, options);
-            const updatedClassDef = await generator.generate();
+            const canonicalGenerator = new CanonicalGenerator(classDef, constants, args.queries.split(','), args.parameter_datasets, options);
+            const updatedClassDef = await canonicalGenerator.generate();
             args.output.end(updatedClassDef.prettyprint());
         }
