+++ tool/autoqa/manual-annotations.js
