+++ tool/lib/bert-canonical-annotator.py
