+++ tool/lib/canonical-extractor.js
